fix wrong assumption that that every newline is a sentence
Originally, for every \r\n in Twitter, we added a '.  ', this assumption is wrong based on the fact that sometimes people use multiple lines. Making message formatting weird, like so: 'Hello.  .  World!" as a substitute for 'Hello\r\n\r\nWorld!'

This removes the dot from the equation just so it looks nicer.